meta: Reorganize file structure
