+++ Includes/xbeScanner.hpp
